+++ test/WebApp.Test/GeneralControllerTest.cs
@@ -1,7 +1,4 @@
-﻿using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Mvc.Testing;
-using Microsoft.AspNetCore.TestHost;
+﻿using Microsoft.AspNetCore.Mvc.Testing;
 using Microsoft.Extensions.DependencyInjection;
 using Moq;
 using NetLah.Extensions.SpaServices.Hosting;
@@ -10,29 +7,6 @@ namespace WebApp.Test;
 
 public class GeneralControllerTest
 {
-    [Theory]
-    [InlineData(false, "http://localhost1/", "http")]
-    [InlineData(true, "http://localhost2/", "https")]
-    public async Task ForwardedHeadersEnabledTest(bool forwardedHeadersEnabled, string url, string scheme)
-    {
-        var builder = TestHelper.CreateWebApplicationBuilder();
-        builder.Configuration["FORWARDEDHEADERS_ENABLED"] = forwardedHeadersEnabled.ToString().ToLower();
-        await using var app = builder.Build();
-
-        app.Run(context =>
-        {
-            Assert.Equal(scheme, context.Request.Scheme);
-            return Task.CompletedTask;
-        });
-
-        await app.StartAsync();
-
-        var client = app.GetTestClient();
-        client.DefaultRequestHeaders.Add("x-forwarded-proto", "https");
-        var result = await client.GetAsync(url);
-        result.EnsureSuccessStatusCode();
-    }
-
     [Fact]
     public async Task GeneralVersionUrl()
     {
